+++ app/Main.hs
@@ -3,7 +3,6 @@ module Main (main) where
 -- entrypoint. It will handle reaping orphans and handling TERM and
 -- INT signals.
 
-import Data.List (foldl')
 import Data.Maybe (fromMaybe)
 import System.Process.PID1
 import System.Environment
@@ -14,18 +13,15 @@ import System.Exit (exitFailure)
 -- | `GetOpt` command line options
 options :: [(String, String)] -> [OptDescr (RunOptions -> RunOptions)]
 options defaultEnv =
-  [ Option ['e'] ["env"] (ReqArg (\opt opts -> setRunEnv (optEnvList (getRunEnv opts) opt) opts) "LIST") "set environment variables from list of comma separated name=value pairs. Can be specified multiple times"
+  [ Option ['e'] ["env"] (ReqArg (\opt opts -> setRunEnv (optEnvList (getRunEnv opts) opt) opts) "ENV") "override environment variable from given name=value pair. Can be specified multiple times to set multiple environment variables"
   , Option ['u'] ["user"] (ReqArg setRunUser "USER") "run command as user"
   , Option ['g'] ["group"] (ReqArg setRunGroup "GROUP") "run command as group"
-  , Option ['w'] ["workdir"] (ReqArg setRunWorkDir "DIR") "command working directory"]
+  , Option ['w'] ["workdir"] (ReqArg setRunWorkDir "DIR") "command working directory"
+  , Option ['t'] ["timeout"] (ReqArg (setRunExitTimeoutSec . read) "TIMEOUT") "timeout (in seconds) to wait for all child processes to exit" ]
   where optEnv env' kv =
           let kvp = fmap (drop 1) $ span (/= '=') kv in
             kvp:filter ((fst kvp /=) . fst) env'
-        split [] = []
-        split s = case fmap (drop 1) $ span (/= ',') s of
-          ("", xs') -> split xs'
-          (x, xs') -> x:split xs'
-        optEnvList env' s = foldl' optEnv (fromMaybe defaultEnv env') $ split s
+        optEnvList = optEnv . fromMaybe defaultEnv
 
 main :: IO ()
 main = do