+++ ChangeLog.md
@@ -1,3 +1,9 @@
+## 0.1.2.0
+
+* Removes support for ',' separated list of environment variables
+  for `-e` command line option
+* Adds support for setting child processes wait timeout on SIGTERM or SIGINT
+
 ## 0.1.1.0
 
 * Adds support for setuid and setguid when running command